refactor: move database types to operations/database
- Relocated database types from src/database/ to src/operations/database/
- Updated all imports to reference new location
- Database types are now alongside query operations (where)
- Exports still available from main index
