--- packages/collections/src/database/index.ts
@@ -1,1 +1,0 @@
-export type { DbAccess, CollectionDbMethods, FindManyQuery, WhereById } from './types'
