+++ packages/collections/src/operations/index.ts
@@ -1,1 +1,2 @@
 export * from './where'
+export * from './database'
